+++ docs/examples/actions/colorizer.py
@@ -7,7 +7,7 @@ class Colorizer(App):
         await self.bind("g", "color('green')")
         await self.bind("b", "color('blue')")
 
-    async def action_color(self, color: str) -> None:
+    def action_color(self, color: str) -> None:
         self.background = f"on {color}"