+++ src/textual/layout_map.py
@@ -4,6 +4,7 @@ from rich.console import Console
 
 from typing import ItemsView, KeysView, ValuesView, NamedTuple
 
+from . import log
 from .geometry import Region, Size
 
 from .widget import Widget
@@ -18,12 +19,7 @@ class RenderRegion(NamedTuple):
 class LayoutMap:
     def __init__(self, size: Size) -> None:
         self.size = size
-        self.contents_region = Region(0, 0, 0, 0)
         self.widgets: dict[Widget, RenderRegion] = {}
-
-    @property
-    def virtual_size(self) -> Size:
-        return self.contents_region.size
 
     def __getitem__(self, widget: Widget) -> RenderRegion:
         return self.widgets[widget]
@@ -42,7 +38,6 @@ class LayoutMap:
 
     def add_widget(
         self,
-        console: Console,
         widget: Widget,
         region: Region,
         order: tuple[int, ...],
@@ -50,16 +45,25 @@ class LayoutMap:
     ) -> None:
         from .view import View
 
-        region += widget.layout_offset
-        self.widgets[widget] = RenderRegion(region, order, clip)
-        self.contents_region = self.contents_region.union(region)
+        if widget in self.widgets:
+            return
+
+        self.widgets[widget] = RenderRegion(region + widget.layout_offset, order, clip)
 
         if isinstance(widget, View):
-            sub_map = widget.layout.generate_map(
-                console, region.size, clip, widget.scroll
-            )
-            widget.virtual_size = sub_map.virtual_size
-            for sub_widget, (sub_region, sub_order, sub_clip) in sub_map.items():
-                sub_region += region.origin
-                sub_clip = sub_clip.intersection(clip)
-                self.add_widget(console, sub_widget, sub_region, sub_order, sub_clip)
+            view: View = widget
+            scroll = view.scroll
+            total_region = region.size.region
+            sub_clip = clip.intersection(region)
+
+            arrangement = view.get_arrangement(region.size, scroll)
+            for sub_region, sub_widget, sub_order in arrangement:
+                total_region = total_region.union(sub_region)
+                if sub_widget is not None:
+                    self.add_widget(
+                        sub_widget,
+                        sub_region + region.origin - scroll,
+                        sub_order,
+                        sub_clip,
+                    )
+            view.virtual_size = total_region.size