+++ src/textual/messages.py
@@ -12,21 +12,32 @@ if TYPE_CHECKING:
 
 
 @rich.repr.auto
-class UpdateMessage(Message, verbosity=3):
-    def __init__(self, sender: MessagePump, widget: Widget, layout: bool = False):
+class Update(Message, verbosity=3):
+    def __init__(self, sender: MessagePump, widget: Widget):
         super().__init__(sender)
         self.widget = widget
-        self.layout = layout
 
     def __rich_repr__(self) -> rich.repr.Result:
         yield self.sender
-        yield "widget"
+        yield self.widget
+
+    def __eq__(self, other: object) -> bool:
+        if isinstance(other, Update):
+            return self.widget == other.widget
+        return NotImplemented
 
     def can_replace(self, message: Message) -> bool:
-        return isinstance(message, UpdateMessage) and self.widget is message.widget
+        return isinstance(message, Update) and self == message
 
 
 @rich.repr.auto
-class LayoutMessage(Message, verbosity=3):
+class Layout(Message, verbosity=3):
     def can_replace(self, message: Message) -> bool:
-        return isinstance(message, LayoutMessage)
+        return isinstance(message, (Layout, Update))
+
+
+@rich.repr.auto
+class CursorMove(Message):
+    def __init__(self, sender: MessagePump, line: int) -> None:
+        self.line = line
+        super().__init__(sender)